Re-doing the big_unsigned interface.
Turns out that git uses two different forms of "big unsigned" or "var
length" encoding integer on it's pack interface.

The first one is on the package header, this is a big-endian formatting
with 3 reserverd bits on the first byte to encode the object type.

The second one is used to encode off-sets inside the packfile, this
format is little-endian. In this case each extra byte is decremented as
to make 0x80 0x01 encode 0x81 :
    0x80 & 0x7f   = 0x00
        0x00 + 1  = 0x01
        0x01 << 7 = 0x80

    0x01 & 0x7f   = 0x7f
      0x80 + 0x01 = 0x81 <- end result

There was also a reorg on the tests for big_unsigned to make them more
sane.