Better stream handling.
devices are factories for buffers and sources
stream factories.

Streams created by the sources have a shared reference
to the source and consequently to the device. When all
the references are gone the device is released.

This scheme makes so each individual stream can have it's
own state without tainting all the others.